RSA
RSA is a acronym for Rivest, Shamir, Adelman, creators of the algorithm. The purpose of RSA is to produce secure, Asymmetric keys based on incredibly large prime numbers. Used in banking, this algorithm becomes incredibly computer intensive when dealing with large primes. 

However, the benefits of RSA is a two definitive key solution in where the 'public' key is used to encrypt and the 'private' to decrypt; where if the public key is leaked, the security of the private key remains secure, hence 'public' key. 

'RSA Keys' illustrates how this algorithm works mathematically, but isn't suitable for real world application due capacity of int.